+++ src/NodeVisitor/OptimizerNodeVisitor.php
@@ -331,7 +331,7 @@ final class OptimizerNodeVisitor extends AbstractNodeVisitor
         }
 
         if ($optimizedChild !== $node->child) {
-            return new GroupNode($optimizedChild, $node->type, $node->name, $node->flags, $node->startPosition, $node->endPosition);
+            return new GroupNode($optimizedChild, $node->type, $node->name, $node->flags, $node->startPosition, $node->endPosition, $node->usePythonSyntax);
         }
 
         return $node;